+++ src/main/java/com/gomo/app/member/infrastructure/EmailAuthSenderService.java
@@ -1,4 +1,4 @@
-package com.gomo.app.auth.infrastructure;
+package com.gomo.app.member.infrastructure;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.mail.javamail.JavaMailSender;
